+++ src/ckmalloc/util.h
@@ -99,8 +99,8 @@ constexpr T CeilDiv(T val, T quotient) {
 // Returns a - b by integer value.
 template <typename T, typename U>
 constexpr size_t PtrDistance(const T* a, const U* b) {
-  return reinterpret_cast<const uint8_t*>(a) -
-         reinterpret_cast<const uint8_t*>(b);
+  return static_cast<size_t>(reinterpret_cast<const uint8_t*>(a) -
+                             reinterpret_cast<const uint8_t*>(b));
 }
 
 // Returns ptr + int, returning void*
